requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.3
python-dotenv
django-cors-headers
djangorestframework-simplejwt
Pillow
